+++ src/components/common/index.ts
@@ -3,4 +3,5 @@ export { default as ClientSwiper } from './ClientSwiper'
 export { default as Modal } from './Modal'
 export { default as MusicSaveBottomSheet } from './MusicSaveBottomSheet'
 export { default as PasswordEditBottomSheet } from './PasswordEditBottomSheet'
+export { default as PlaylistBottomSheet } from './PlaylistBottomSheet'
 export { default as ProfileEditBottomSheet } from './ProfileEditBottomSheet'
